Fixed compatibility errors with user, communitygroup, and resource classes

+++ project/src/main/java/com/example/javadb/repository/CommunityGroupRepository.java
@@ -1,8 +1,11 @@
+/**
+ * This package contains repositories for managing Community Groups in the application.
+ */
 package com.example.javadb.repository;
 
 import com.example.javadb.model.CommunityGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommunityGroupRepository extends JpaRepository<CommunityGroup, Long> {
+public interface CommunityGroupRepository extends JpaRepository<CommunityGroup, Integer> {
 
 }
