+++ project/src/main/java/com/example/javadb/repository/UserRepository.java
@@ -1,8 +1,11 @@
+/**
+ * This package contains repositories for managing Users in the application.
+ */
 package com.example.javadb.repository;
 
 import com.example.javadb.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
 }
